+++ src/api/projects.lisp
@@ -8,8 +8,10 @@
   (:import-from #:alexandria
                 #:plist-hash-table)
   (:import-from #:serapeum
+                #:fmt
                 #:dict)
   (:import-from #:ultralisp/models/project
+                #:get-project2-by-id
                 #:project2)
   (:import-from #:mito
                 #:object-id)
@@ -25,7 +27,7 @@
   (:param limit integer "Maximum number of projects per page.")
   (:result (paginated-list-of project2))
   
-  (log:error "Retrieving projects with" tag page-key)
+  (log:info "Retrieving projects with" tag page-key)
   (ultralisp/db:with-connection ()
     (ultralisp/models/tag::get-projects-by-tag-paginated tag
                                                          :page-key page-key
@@ -38,13 +40,36 @@
   (:result (list-of string))
   
   (ultralisp/db:with-connection ()
-    (ultralisp/models/tag::get-project-tags project-id)))
+    (let ((project (ultralisp/models/project:get-project2-by-id project-id)))
+      (unless project
+        (openrpc-server:return-error (fmt "Project with id ~A not found."
+                                          project-id)
+                                     :code -1))
+      (ultralisp/models/tag::get-project-tags project))))
 
 
 (define-rpc-method (api get-project-by-name) (name)
   (:summary "Returns a project details by it's name. Name should be in it's full form like \"40ants/doc\".")
   (:param name string)
-  (:result (or null
-               ultralisp/models/project:project2))
+  (:result ultralisp/models/project:project2)
   (ultralisp/db:with-connection ()
-    (ultralisp/models/project:get-project2 name)))
+    (let ((project (ultralisp/models/project:get-project2 name)))
+      (unless project
+        (openrpc-server:return-error (fmt "Project with name ~S not found."
+                                          name)
+                                     :code -2))
+      (values project))))
+
+
+(define-rpc-method (api get-project-systems) (project-id)
+  (:summary "Retrieve all systems of a given project.")
+  (:param project-id integer "ID of a project.")
+  (:result (list-of quickdist:system-info))
+  
+  (ultralisp/db:with-connection ()
+    (let ((project (get-project2-by-id project-id)))
+      (unless project
+        (openrpc-server:return-error (fmt "Project with id ~A not found."
+                                          project-id)
+                                     :code -1))
+      (ultralisp/models/project::get-project-systems project))))