+++ src/api/server.lisp
@@ -4,10 +4,17 @@
   (:import-from #:openrpc-server
                 #:type-to-schema)
   (:import-from #:ultralisp/api/projects)
-  (:import-from #:ultralisp/api/search))
+  (:import-from #:ultralisp/api/search)
+  (:import-from #:ultralisp/api/tags))
 (in-package #:ultralisp/api/server)
 
 
-(defmethod openrpc-server/interface:slots-to-exclude ((type mito.dao.table:dao-table-class))
+(defmethod openrpc-server/interface:slots-to-exclude ((obj mito.dao:dao-table-class))
   (list* "synced"
+         (call-next-method)))
+
+
+(defmethod openrpc-server/interface:slots-to-exclude ((type (eql (find-class 'quickdist:system-info))))
+  (list* "path"
+         "project-name"
          (call-next-method)))